+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 
+
 ## v4.22.0
+  - [EE] new formatter `{d.pdf_url:appendFile}` to append one or several PDFs at the end of the generated document. 
+    This formatter accepts either a single URL or an array of URLs. It can be positioned anywhere within a template and does not produce any text output by itself.
+    Carbone will return an error and will not generate the report in the following scenarios:
+      - If the provided URL returns an error (with 2 retries for Carbone Cloud).
+      - If the downloaded document is not recognized as a valid PDF.
+      - If more than 20 files are to be downloaded (On-Premise parameter: `maxDownloadFileCount`).
+      - If the total size of all downloaded files exceeds 10 MB (On-Premise parameter: `maxDownloadFileSizeTotal`).
+    Carbone will ignore the formatter and proceed to generate the report in the following cases:
+      - If the provided URL is null, undefined, or an empty string.
+      - If the final report is not in PDF format
+  - Improve file type detection when inserting images or PDF coming from external URL. It can read file extension in Content-Disposition if Carbone cannot find it Content-Type and URL.
   - The execution order of multiple Carbone tags using the `:set` formatter is guaranteed within the same part of a document (body, header, footer, text box).
     As a result, you can create new variables that depend on previously created ones. Example:
     ```